--- tests/property_test/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "property_test"
-description = "Bindgen property tests. Generate random valid C code and pass it to the csmith/predicate.py script"
-version = "0.1.0"
-authors = ["Shea Newton <snewton@polysync.io>"]
-
-[dependencies]
-quickcheck = "0.4"
-tempdir = "0.3"
-rand = "0.3"
