--- tests/property_test/src/lib.rs
@@ -1,5 +1,0 @@
-extern crate quickcheck;
-extern crate rand;
-extern crate tempdir;
-
-pub mod fuzzers;
